+++ CHANGELOG.md
@@ -35,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_blender.utilities.draw_curves`.
 * Added `compas_blender.utilities.draw_surfaces`.
 * Added `rgba` and `rgba255` properties to `compas.colors.Color`.
-
+* Added `from_html_name` method to `compas.colors.Color`.
 ### Changed
 
 * Fixed bug in `mesh_slice_plane()` , `Mesh.slice_plane()`.